refactor(engine): move the custody-columns helper to the shared partial
Review follow-up. The helper is called from the Amsterdam and Bogota
partials and its tracker field is declared in the cross-fork file, so that
is where it belongs now that it is shared.
